Update package.json
The prior author hasn't made commits in a while, and the repo seems to be abandoned. Thus, I have forked the repo.

I added Juan Leon as a contributor and myself as the author. I found no protocol for repo forks and authors despite search and polling. Commit history and "forked from" show his original contributions, without which this package wouldn't be around.